added query builder (filtering & sorting)

+++ service/src/main/java/com/epam/esm/service/impl/GiftCertificateServiceImpl.java
@@ -7,16 +7,15 @@ import com.epam.esm.mapper.impl.GiftCertificateMapper;
 import com.epam.esm.repository.CertificatePurchaseRepository;
 import com.epam.esm.repository.GiftCertificateRepository;
 import com.epam.esm.service.GiftCertificateService;
+import com.epam.esm.util.CertificateDateFormatter;
 import com.epam.esm.validator.GiftCertificateValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
+import java.time.LocalDateTime;
+import java.util.*;
 import java.util.stream.Collectors;
 
 import static com.epam.esm.util.ParameterName.*;
@@ -27,16 +26,19 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     private final CertificatePurchaseRepository purchaseRepository;
     private final GiftCertificateValidator validator;
     private final GiftCertificateMapper mapper;
+    private final CertificateDateFormatter dateFormatter;
 
     @Autowired
     public GiftCertificateServiceImpl(GiftCertificateRepository repository,
                                       CertificatePurchaseRepository purchaseRepository,
                                       GiftCertificateValidator validator,
-                                      @Qualifier("certificateMapper") GiftCertificateMapper mapper) {
+                                      @Qualifier("certificateMapper") GiftCertificateMapper mapper,
+                                      CertificateDateFormatter dateFormatter) {
         this.repository = repository;
         this.purchaseRepository = purchaseRepository;
         this.validator = validator;
         this.mapper = mapper;
+        this.dateFormatter = dateFormatter;
     }
 
     @Override
@@ -61,21 +63,10 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public boolean update(Map<String, Object> certificateData) {
-        if (certificateData.containsKey(ID) && validator.checkCertificate(certificateData)) {
-            GiftCertificate giftCertificate = new GiftCertificate();
+        if (!certificateData.isEmpty() && certificateData.containsKey(ID)
+                && validator.checkCertificate(certificateData)) {
+            GiftCertificate giftCertificate = retrieveCertificateData(certificateData);
             giftCertificate.setId(Long.parseLong(String.valueOf(certificateData.get(ID))));
-            if (certificateData.containsKey(NAME)) {
-                giftCertificate.setName(String.valueOf(certificateData.get(NAME)));
-            }
-            if (certificateData.containsKey(DESCRIPTION)) {
-                giftCertificate.setDescription(String.valueOf(certificateData.get(DESCRIPTION)));
-            }
-            if (certificateData.containsKey(PRICE)) {
-                giftCertificate.setPrice(new BigDecimal(String.valueOf(certificateData.get(PRICE))));
-            }
-            if (certificateData.containsKey(DURATION)) {
-                giftCertificate.setDuration(Integer.parseInt(String.valueOf(certificateData.get(DURATION))));
-            }
             if (certificateData.containsKey(TAGS)) {
                 Set<Tag> tags = ((ArrayList<Map<String, String>>) certificateData.get(TAGS)).stream()
                         .map(t -> t.get(NAME))
@@ -91,7 +82,12 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public boolean delete(long id) {
-        return repository.delete(id);
+        if (repository.findById(id).isPresent()) {
+            repository.delete(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
@@ -114,18 +110,43 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     }
 
     @Override
-    public Set<GiftCertificateDto> findByName(String name) {
-        Set<GiftCertificate> certificates = repository.findByName(name);
-        return certificates.stream()
-                .map(mapper::mapToDto)
-                .collect(Collectors.toSet());
+    public Set<GiftCertificateDto> findBySeveralParameters(Map<String, Object> certificateData, List<String> sortTypes) {
+        if (!certificateData.isEmpty() && validator.checkCertificate(certificateData)) {
+            GiftCertificate giftCertificate = retrieveCertificateData(certificateData);
+            String tagName = null;
+            if (certificateData.containsKey(TAG)) {
+                tagName = String.valueOf(certificateData.get(TAG));
+            }
+            Set<GiftCertificate> certificates = repository.findBySeveralParameters(giftCertificate, tagName, sortTypes);
+            return certificates.stream()
+                    .map(mapper::mapToDto)
+                    .collect(Collectors.toCollection(LinkedHashSet::new));
+        }
+        return new LinkedHashSet<>();
     }
 
-    @Override
-    public Set<GiftCertificateDto> findByDescription(String description) {
-        Set<GiftCertificate> certificates = repository.findByDescription(description);
-        return certificates.stream()
-                .map(mapper::mapToDto)
-                .collect(Collectors.toSet());
+    public GiftCertificate retrieveCertificateData(Map<String, Object> certificateData) {
+        GiftCertificate giftCertificate = new GiftCertificate();
+        if (certificateData.containsKey(NAME)) {
+            giftCertificate.setName(String.valueOf(certificateData.get(NAME)));
+        }
+        if (certificateData.containsKey(DESCRIPTION)) {
+            giftCertificate.setDescription(String.valueOf(certificateData.get(DESCRIPTION)));
+        }
+        if (certificateData.containsKey(PRICE)) {
+            giftCertificate.setPrice(new BigDecimal(String.valueOf(certificateData.get(PRICE))));
+        }
+        if (certificateData.containsKey(DURATION)) {
+            giftCertificate.setDuration(Integer.parseInt(String.valueOf(certificateData.get(DURATION))));
+        }
+        if (certificateData.containsKey(CREATE_DATE)) {
+            LocalDateTime formattedDate = dateFormatter.format(String.valueOf(certificateData.get(CREATE_DATE)));
+            giftCertificate.setCreateDate(formattedDate);
+        }
+        if (certificateData.containsKey(LAST_UPDATE_DATE)) {
+            LocalDateTime formattedDate = dateFormatter.format(String.valueOf(certificateData.get(LAST_UPDATE_DATE)));
+            giftCertificate.setLastUpdateDate(formattedDate);
+        }
+        return giftCertificate;
     }
 }
